Backend/requirements.txt
apscheduler==3.10.4
Flask==3.0.3
Flask_Cors==5.0.0
groq==0.11.0
langchain==0.3.4
langchain_community==0.3.3
langchain_google_genai==2.0.1
langchain_groq==0.2.0
mail==2.1.0
Markdown==3.5.2
Pillow==11.0.0
protobuf==5.28.2
psycopg2==2.9.9
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
pytube==15.0.0
spacy==3.7.6
streamlit==1.32.2
youtube_transcript_api==0.6.2
redis===5.1.1
yt_dlp
assemblyai===0.34.0
movie
